refactor: 데이터베이스 하나로 통일
