+++ app/src/androidTest/java/woowacourse/movie/database/MovieDatabaseTest.kt
@@ -5,14 +5,22 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.After
 import org.junit.Before
 import org.junit.Test
+import woowacourse.movie.data.dao.MovieDao
 import woowacourse.movie.data.dao.MovieReservationDao
+import woowacourse.movie.data.dao.MovieTheaterDao
 import woowacourse.movie.data.database.MovieReservationDatabase
+import woowacourse.movie.data.entity.Movie
 import woowacourse.movie.data.entity.MovieReservation
+import woowacourse.movie.data.entity.MovieTheater
 import woowacourse.movie.fixtures.context
 import java.io.IOException
 
-class MovieReservationTest {
+class MovieDatabaseTest {
     private lateinit var movieReservationDao: MovieReservationDao
+    private lateinit var movieDao: MovieDao
+
+    private lateinit var theaterDao: MovieTheaterDao
+
     private lateinit var db: MovieReservationDatabase
 
     @Before
@@ -21,7 +29,12 @@ class MovieReservationTest {
             Room.inMemoryDatabaseBuilder(
                 context, MovieReservationDatabase::class.java,
             ).build()
+        movieDao = db.movieDao()
+        theaterDao = db.movieTheaterDao()
         movieReservationDao = db.movieReservationDao()
+
+        movieDao.insert(Movie.STUB)
+        theaterDao.insertAll(MovieTheater.STUB_A, MovieTheater.STUB_B, MovieTheater.STUB_C)
     }
 
     @After
