--- app/src/androidTest/java/woowacourse/movie/database/MovieTest.kt
@@ -1,40 +1,0 @@
-package woowacourse.movie.database
-
-import androidx.room.Room
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.After
-import org.junit.Before
-import org.junit.Test
-import woowacourse.movie.data.dao.MovieDao
-import woowacourse.movie.data.database.MovieDatabase
-import woowacourse.movie.data.entity.Movie
-import woowacourse.movie.fixtures.context
-import java.io.IOException
-
-class MovieTest {
-    private lateinit var movieDao: MovieDao
-    private lateinit var db: MovieDatabase
-
-    @Before
-    fun setUp() {
-        db =
-            Room.inMemoryDatabaseBuilder(
-                context, MovieDatabase::class.java,
-            ).build()
-        movieDao = db.movieDao()
-    }
-
-    @After
-    @Throws(IOException::class)
-    fun closeDb() {
-        db.close()
-    }
-
-    @Test
-    fun insertMovies() {
-        movieDao.insertAll(Movie.STUB)
-
-        val movie = movieDao.getAll()
-        assertThat(movie).contains(Movie.STUB)
-    }
-}
