--- app/src/main/java/woowacourse/movie/data/database/MovieDatabase.kt
@@ -1,14 +1,0 @@
-package woowacourse.movie.data.database
-
-import androidx.room.Database
-import androidx.room.RoomDatabase
-import androidx.room.TypeConverters
-import woowacourse.movie.data.converter.LocalDateTimeConverters
-import woowacourse.movie.data.dao.MovieDao
-import woowacourse.movie.data.entity.Movie
-
-@Database(entities = [Movie::class], version = 1)
-@TypeConverters(LocalDateTimeConverters::class)
-abstract class MovieDatabase : RoomDatabase() {
-    abstract fun movieDao(): MovieDao
-}
